rename "boughtList"->"shoppingList" add data.
