+++ RaiseTech08/src/main/resources/application.properties
@@ -13,5 +13,5 @@ spring.datasource.password=
 #spring.h2.console.enabled=true
 
 # mybatis
-mybatis.mapper-locations=classpath*:/mapper/boughtMapper.xml
+mybatis.mapper-locations=classpath*:/mapper/ShoppingMapper.xml
 mybatis.configuration.map-underscore-to-camel-case=true
