+++ RaiseTech08/src/main/resources/mapper/ShoppingMapper.xml
@@ -3,10 +3,10 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <!-- Mppperとxmlのマッピング -->
-<mapper namespace="com.example.demo.Repository.BoughtMapper">
+<mapper namespace="com.example.demo.Repository.ShoppingMapper">
 	<!-- 全件取得 -->
 	<select id="selectAll"
-		resultType="com.example.demo.Form.BoughtListForm">
-		SELECT * FROM BoughtList
+		resultType="com.example.demo.entity.ShoppingList">
+		SELECT * FROM ShoppingList
 	</select>
 </mapper>
