+++ RaiseTech08/src/test/resources/data.sql
@@ -1,4 +1,4 @@
-INSERT INTO `BoughtList` (
+INSERT INTO `ShoppingList` (
 	`id`,
 	`item_name`,
 	`item_count`,
@@ -11,4 +11,16 @@ INSERT INTO `BoughtList` (
 	200,
 	'美味しいよ'
 	);
-	
+INSERT INTO `ShoppingList` (
+	`id`,
+	`item_name`,
+	`item_count`,
+	`item_price`,
+	`item_memo`
+) VALUES (
+	2,
+	'オレンジ',
+	1,
+	300,
+	'酸っぱいよ'
+	);
