+++ RaiseTech08/src/test/resources/schema.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS BoughtList(
+CREATE TABLE IF NOT EXISTS ShoppingList(
 	id INT AUTO_INCREMENT NOT NULL PRIMARY KEY,
 	item_name VARCHAR(10),
 	item_count int,
