+++ RaiseTech08/src/test/resources/test.properties
@@ -13,7 +13,7 @@ spring.datasource.password=
 #spring.h2.console.enabled=true
 
 # mybatis
-mybatis.mapper-locations=classpath*:/mapper/boughtMapper.xml
+mybatis.mapper-locations=classpath*:/mapper/ShoppingMapper.xml
 mybatis.configuration.map-underscore-to-camel-case=true
 
 spring.mvc.hiddenmethod.filter.enabled=true
